move some hotel tests
